+++ client-service/src/main/java/com/jaspersoft/android/sdk/service/server/JasperServer.java
@@ -1,9 +1,9 @@
-package com.jaspersoft.android.sdk.service;
+package com.jaspersoft.android.sdk.service.server;
 
 import com.jaspersoft.android.sdk.network.api.ServerRestApi;
 import com.jaspersoft.android.sdk.service.data.server.ServerInfo;
-import com.jaspersoft.android.sdk.service.result.ContentResult;
-import com.jaspersoft.android.sdk.service.result.PendingResult;
+import com.jaspersoft.android.sdk.service.common.ContentResult;
+import com.jaspersoft.android.sdk.service.common.PendingResult;
 
 /**
  * @author Tom Koptel
