+++ client-service/src/test/java/com/jaspersoft/android/sdk/service/server/JasperServerImplTest.java
@@ -1,4 +1,4 @@
-package com.jaspersoft.android.sdk.service;
+package com.jaspersoft.android.sdk.service.server;
 
 import com.jaspersoft.android.sdk.network.api.ServerRestApi;
 import com.jaspersoft.android.sdk.network.entity.server.ServerInfoResponse;
